# frontend-integration-package/openapi.snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlowLedger API",
    "version": "0.1.0",
    "description": "API for managing clients, audits, industries, and related business processes"
  },
...
  "paths": {
    "/api/ai/chat": {
      "post": {
        "summary": "Chat completion",
        "tags": [
          "AI"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "messages"
                ],
                "properties": {
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": [
                            "system",
                            "user",
                            "assistant"
                          ]
                        },
                        "content": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "model": {
                    "type": "string"
                  },
                  "temperature": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assistant reply"
          }
        }
      }
    },
    "/api/ai/chat-stream": {
      "post": {
        "summary": "Chat completion (SSE stream)",
        "tags": [
          "AI"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "messages"
                ],
                "properties": {
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": [
                            "system",
                            "user",
                            "assistant"
                          ]
                        },
                        "content": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "model": {
                    "type": "string"
                  },
                  "temperature": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server-sent events stream"
          }
        }
      }
    },
    "/api/ai/assist": {
      "post": {
        "summary": "Tool-enabled assistant that can fetch server data",
        "tags": [
          "AI"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "messages"
                ],
                "properties": {
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": [
                            "system",
                            "user",
                            "assistant"
                          ]
                        },
                        "content": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "model": {
                    "type": "string"
                  },
                  "temperature": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assistant reply with tool trace"
          }
        }
      }
    },
    "/api/ai/tag-suggest": {
      "post": {
        "summary": "Suggest tags for a note (reuse existing, propose few new)",
        "tags": [
          "AI"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "client_id",
                  "note"
                ],
                "properties": {
                  "client_id": {
                    "type": "integer"
                  },
                  "note": {
                    "type": "string"
                  },
                  "maxExisting": {
                    "type": "integer",
                    "description": "Max existing tags to suggest"
                  },
                  "maxNew": {
                    "type": "integer",
                    "description": "Max new tags to propose"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "existing": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "tag_id": {
                                "type": "integer"
                              },
                              "tag_name": {
                                "type": "string"
                              },
                              "reason": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "new": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "tag_name": {
                                "type": "string"
                              },
                              "reason": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "rationale": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/name-suggest": {
      "post": {
        "summary": "Suggest names for an engagement",
        "tags": [
          "AI"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "current_name"
                ],
                "properties": {
                  "client_id": {
                    "type": "integer"
                  },
                  "current_name": {
                    "type": "string"
                  },
                  "context": {
                    "type": "string"
                  },
                  "maxSuggestions": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Name suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}